+++ .github/workflows/rust.yml
@@ -12,7 +12,7 @@ on:
 env:
   SHA: ${{ github.event.pull_request.head.sha || github.sha }}
   CARGO_TERM_COLOR: always
-  RUST_VERSION: "1.93.1"
+  RUST_VERSION: "1.95.0"
 
 jobs:
   check:
