+++ Cargo.toml
@@ -17,7 +17,7 @@ spz-lib = { git = "https://github.com/404-Repo/spz-rs.git", rev = "227c6cb6cdc25
     "async",
 ] }
 multer = "3.1.0"
-tokio = { version = "1.52.0", features = [
+tokio = { version = "1.52.1", features = [
     "rt-multi-thread",
     "macros",
     "sync",
