+++ src/http3/handlers/common/mod.rs
@@ -2,4 +2,5 @@ pub mod activity;
 pub mod model_errors;
 pub mod multipart;
 pub mod origin;
+pub mod peer;
 pub mod worker_auth;
